+++ src/main/java/Ken/exception/KenException.java
@@ -1,4 +1,4 @@
-package Ken;
+package ken.exception;
 
 public class KenException extends Exception {
     public KenException(String message) {
