+++ src/main/java/Ken/task/Event.java
@@ -1,4 +1,4 @@
-package Ken.task;
+package ken.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
